moving types to shared types file
