+++ packages/gamut-labs/src/Tabs/TabList.tsx
@@ -1,12 +1,9 @@
+import { Box } from '@codecademy/gamut';
 import React, { ReactElement } from 'react';
 
 import { Tab } from './Tab';
 import { TabPanel, TabPanelProps } from './TabPanel';
-
-type TabListProps = Pick<TabsProps, 'children'> & {
-  activeTabIndex: number;
-  setActiveTab: (index: number) => void;
-};
+import { TabListProps } from './types';
 
 export const TabList: React.FC<TabListProps> = ({
   children,
@@ -14,20 +11,24 @@ export const TabList: React.FC<TabListProps> = ({
   setActiveTab,
 }) => {
   const childArray = React.Children.toArray(children) as ReactElement[];
-  return childArray
-    .filter((c) => c.type === TabPanel)
-    .map((ChildTab, i) => {
-      const { title, onTabClick } = ChildTab.props as TabPanelProps;
-      const onClick = () => {
-        onTabClick?.();
-        setActiveTab(i);
-      };
-      return (
-        <Tab
-          title={title}
-          isActiveTab={i === activeTabIndex}
-          onClick={onClick}
-        />
-      );
-    });
+  return (
+    <Box role="tablist">
+      {childArray
+        .filter((c) => c.type === TabPanel)
+        .map((ChildTab, i) => {
+          const { title, onTabClick } = ChildTab.props as TabPanelProps;
+          const onClick = () => {
+            onTabClick?.();
+            setActiveTab(i);
+          };
+          return (
+            <Tab
+              title={title}
+              isActiveTab={i === activeTabIndex}
+              onClick={onClick}
+            />
+          );
+        })}
+    </Box>
+  );
 };
